backend/requirements.txt
alembic==1.16.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-Dance==7.1.0
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
URLObject==2.4.3
Werkzeug==3.1.3
